chore: move validate package from pkg to internal
