+++ cli/internal/validate/refs.go
