+++ cli/internal/validate/refs_test.go
