+++ cli/internal/validate/required_keys.go
