+++ cli/internal/validate/required_keys_test.go
