+++ cli/internal/validate/validate_test.go
